# number-generate-service/src/main/java/com/itqtest/numbersgenerateservice/exception/handler/ErrorResponseDto.java
package com.itqtest.numbersgenerateservice.exception.handler;

/**
 * DTO для представления сообщения об ошибке.
 * Используется для передачи сообщений об ошибках от сервисов или контроллеров
 *
 * @author Egor Nazarev
 */
public record ErrorResponseDto(String errorMessage) {}