# orders-service/src/main/java/com/itqtest/ordersservice/exception/handler/ErrorResponseDto.java
package com.itqtest.ordersservice.exception.handler;

/**
 * DTO для представления сообщения об ошибке.
 * Используется для передачи сообщений об ошибках от сервисов или контроллеров
 *
 * @author Egor Nazarev
 */
public record ErrorResponseDto(String errorMessage) {}